NewLangHelperWeb/requirements.txt
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-cors-headers==2.0.2
django-oauth-toolkit==0.12.0
django-pipeline==1.6.12
django-rest-auth==0.9.1
djangorestframework==3.6.2
gunicorn==19.7.1
numpy==1.12.1
oauthlib==2.0.1
psycopg2==2.7.1
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-django==1.2.0
social-auth-core==1.3.0
